#110 Rename mapping rule classes according to their semantics
