+++ conformance/src/test/scala/za/co/absa/enceladus/conformance/interpreter/rules/RulesSuite.scala
@@ -33,28 +33,28 @@ class RulesSuite extends FunSuite with SparkTestBase {
 
   test("Test country code join condition") {
     val countryRule = EmployeeConformance.countryRule
-    val countryCondGen = MappingRuleInterpreter.getJoinCondition(countryRule).expr
+    val countryCondGen = MappingRuleInterpreterNoExplode.getJoinCondition(countryRule).expr
     val countryCond = (lit(true)
-      && (col(s"${MappingRuleInterpreter.inputDfAlias}.country") === col(s"${MappingRuleInterpreter.mappingTableAlias}.country_code"))) expr
+      && (col(s"${MappingRuleInterpreterNoExplode.inputDfAlias}.country") === col(s"${MappingRuleInterpreterNoExplode.mappingTableAlias}.country_code"))) expr
 
     assert(countryCondGen.semanticEquals(countryCond))
   }
 
   test("Test department join condition") {
     val deptRule = EmployeeConformance.departmentRule
-    val deptCondGen = MappingRuleInterpreter.getJoinCondition(deptRule).expr
+    val deptCondGen = MappingRuleInterpreterNoExplode.getJoinCondition(deptRule).expr
     val deptCond = (lit(true) &&
-      (col(s"${MappingRuleInterpreter.inputDfAlias}.dept") === col(s"${MappingRuleInterpreter.mappingTableAlias}.dept_id"))) expr
+      (col(s"${MappingRuleInterpreterNoExplode.inputDfAlias}.dept") === col(s"${MappingRuleInterpreterNoExplode.mappingTableAlias}.dept_id"))) expr
 
     assert(deptCondGen.semanticEquals(deptCond))
   }
 
   test("Test role join condition") {
     val roleRule = EmployeeConformance.roleRule
-    val roleCondGen = MappingRuleInterpreter.getJoinCondition(roleRule).expr
+    val roleCondGen = MappingRuleInterpreterNoExplode.getJoinCondition(roleRule).expr
     val roleCond = (lit(true) &&
-      (col(s"${MappingRuleInterpreter.inputDfAlias}.role") <=> col(s"${MappingRuleInterpreter.mappingTableAlias}.role_id")) &&
-      (col(s"${MappingRuleInterpreter.inputDfAlias}.conformed_country") <=> col(s"${MappingRuleInterpreter.mappingTableAlias}.country"))) expr
+      (col(s"${MappingRuleInterpreterNoExplode.inputDfAlias}.role") <=> col(s"${MappingRuleInterpreterNoExplode.mappingTableAlias}.role_id")) &&
+      (col(s"${MappingRuleInterpreterNoExplode.inputDfAlias}.conformed_country") <=> col(s"${MappingRuleInterpreterNoExplode.mappingTableAlias}.country"))) expr
 
     assert(roleCondGen.semanticEquals(roleCond))
   }
@@ -111,48 +111,48 @@ class RulesSuite extends FunSuite with SparkTestBase {
         )
       ))
 
-    MappingRuleInterpreter.ensureDefaultValueMatchSchema("id_test", schema, "id", "1")
-    MappingRuleInterpreter.ensureDefaultValueMatchSchema("name_test", schema, "name", "'test'")
-    MappingRuleInterpreter.ensureDefaultValueMatchSchema("name_null_ok", schema, "name", "null")
-    MappingRuleInterpreter.ensureDefaultValueMatchSchema("decimal_test", schema, "price", "1.6127")
-    MappingRuleInterpreter.ensureDefaultValueMatchSchema("date_test", schema, "orders.orderdate", "'2017-10-25'")
-    MappingRuleInterpreter.ensureDefaultValueMatchSchema("timestamp_test", schema, "orders.delivertime", "'2017-10-25 08:35:43'")
-    MappingRuleInterpreter.ensureDefaultValueMatchSchema("boolean_test", schema, "orders.happy", "true")
-    MappingRuleInterpreter.ensureDefaultValueMatchSchema("struct_test", schema, "orders.system", "struct('Unknown' as name, 'None' as description)")
+    MappingRuleInterpreterNoExplode.ensureDefaultValueMatchSchema("id_test", schema, "id", "1")
+    MappingRuleInterpreterNoExplode.ensureDefaultValueMatchSchema("name_test", schema, "name", "'test'")
+    MappingRuleInterpreterNoExplode.ensureDefaultValueMatchSchema("name_null_ok", schema, "name", "null")
+    MappingRuleInterpreterNoExplode.ensureDefaultValueMatchSchema("decimal_test", schema, "price", "1.6127")
+    MappingRuleInterpreterNoExplode.ensureDefaultValueMatchSchema("date_test", schema, "orders.orderdate", "'2017-10-25'")
+    MappingRuleInterpreterNoExplode.ensureDefaultValueMatchSchema("timestamp_test", schema, "orders.delivertime", "'2017-10-25 08:35:43'")
+    MappingRuleInterpreterNoExplode.ensureDefaultValueMatchSchema("boolean_test", schema, "orders.happy", "true")
+    MappingRuleInterpreterNoExplode.ensureDefaultValueMatchSchema("struct_test", schema, "orders.system", "struct('Unknown' as name, 'None' as description)")
 
     assert(intercept[ValidationException] {
-      MappingRuleInterpreter.ensureDefaultValueMatchSchema("no_attribute_test", schema, "code", "")
+      MappingRuleInterpreterNoExplode.ensureDefaultValueMatchSchema("no_attribute_test", schema, "code", "")
     }.getMessage contains "does not contain the specified target attribute")
 
     assert(intercept[ValidationException] {
-      MappingRuleInterpreter.ensureDefaultValueMatchSchema("not_string", schema, "name", "struct('Unknown' as name)")
+      MappingRuleInterpreterNoExplode.ensureDefaultValueMatchSchema("not_string", schema, "name", "struct('Unknown' as name)")
     }.getMessage contains "A string expected")
 
     println(intercept[ValidationException] {
-      MappingRuleInterpreter.ensureDefaultValueMatchSchema("id_null", schema, "id", "null")
+      MappingRuleInterpreterNoExplode.ensureDefaultValueMatchSchema("id_null", schema, "id", "null")
     }.getMessage)
 
     intercept[ValidationException] {
-      MappingRuleInterpreter.ensureDefaultValueMatchSchema("id_test", schema, "id", "wrong")
+      MappingRuleInterpreterNoExplode.ensureDefaultValueMatchSchema("id_test", schema, "id", "wrong")
     }
 
     assert(intercept[ValidationException] {
-      MappingRuleInterpreter.ensureDefaultValueMatchSchema("decimal_test", schema, "price", "12345.67")
+      MappingRuleInterpreterNoExplode.ensureDefaultValueMatchSchema("decimal_test", schema, "price", "12345.67")
     }.getMessage contains "Scale/precision don't match the value")
 
     assert(intercept[ValidationException] {
-      MappingRuleInterpreter.ensureDefaultValueMatchSchema("date_test", schema, "orders.orderdate", "'25/10/2017'")
+      MappingRuleInterpreterNoExplode.ensureDefaultValueMatchSchema("date_test", schema, "orders.orderdate", "'25/10/2017'")
     }.getMessage contains "Make sure the value matches 'yyyy-MM-dd'")
 
     assert(intercept[ValidationException] {
-      MappingRuleInterpreter.ensureDefaultValueMatchSchema("timestamp_test", schema, "orders.delivertime", "'25-10-201708:25:43'")
+      MappingRuleInterpreterNoExplode.ensureDefaultValueMatchSchema("timestamp_test", schema, "orders.delivertime", "'25-10-201708:25:43'")
     }.getMessage contains "Make sure the value matches 'yyyy-MM-dd HH:mm:ss'")
 
     intercept[ValidationException] {
-      MappingRuleInterpreter.ensureDefaultValueMatchSchema("boolean_test", schema, "orders.happy", "a")
+      MappingRuleInterpreterNoExplode.ensureDefaultValueMatchSchema("boolean_test", schema, "orders.happy", "a")
     }
     intercept[ValidationException] {
-      MappingRuleInterpreter.ensureDefaultValueMatchSchema("struct_test", schema, "orders.system", "struct('unknown' as name)")
+      MappingRuleInterpreterNoExplode.ensureDefaultValueMatchSchema("struct_test", schema, "orders.system", "struct('unknown' as name)")
     }
 
   }
